terraform/cf_datapipeline/requirements.txt
google-cloud-bigquery
google-cloud-storage
csvfile
urllib3